# Conflicts:
#	backend/src/main/java/ase/meditrack/controller/TeamController.java
#	backend/src/main/java/ase/meditrack/controller/UserController.java
#	backend/src/main/java/ase/meditrack/model/TeamValidator.java
#	backend/src/main/java/ase/meditrack/model/UserValidator.java
#	backend/src/main/java/ase/meditrack/model/dto/UserDto.java
#	backend/src/main/java/ase/meditrack/model/mapper/RoleMapper.java
#	backend/src/main/java/ase/meditrack/service/TeamService.java
#	backend/src/main/java/ase/meditrack/service/UserService.java
#	frontend/src/app/app.module.ts

+++ frontend/src/app/components/roles/roles.component.ts
@@ -28,13 +28,16 @@ export class RolesComponent {
   }
 
   deleteRole(role: Role): void {
-    this.rolesService.deleteRole(role.id)
-      .subscribe(response => {
-        console.log('Role deleted successfully:', response);
-        this.loadRoles();
-      }, error => {
-        console.error('Error deleting role:', error);
-      });
+    if (role.id != null) {
+      this.rolesService.deleteRole(role.id)
+        .subscribe({
+          next: (response) => {
+          console.log('Role deleted successfully:', response);
+          this.loadRoles();
+        }, error: (error) => {
+          console.error('Error deleting role:', error);
+        }});
+    }
   }
 
   createRole() {
@@ -43,13 +46,14 @@ export class RolesComponent {
         name: this.newRoleName
       };
       this.rolesService.createRole(newRole)
-        .subscribe(response => {
+        .subscribe({
+          next: (response) => {
           console.log('Role created successfully:', response);
           this.loadRoles();
           this.newRoleName = '';
-        }, error => {
+        }, error: (error) => {
           console.error('Error creating role:', error);
-        });
+      }});
       this.showNewRoleInputField = false;
     } else {
       console.error('Role name must be unique.');
